Some minor work on collapse mechanics:  - spikes have special collapse handling and can fall together (which just looks awesome)  - explosions now directly trigger collapses.

+++ src/main/java/net/dries007/tfc/world/tracker/CollapseData.java
@@ -14,7 +14,7 @@ public class CollapseData implements INBTSerializable<CompoundNBT>
     List<BlockPos> nextPositions;
     double radiusSquared;
 
-    public CollapseData(BlockPos centerPos, List<BlockPos> nextPositions, int radiusSquared)
+    public CollapseData(BlockPos centerPos, List<BlockPos> nextPositions, double radiusSquared)
     {
         this.centerPos = centerPos;
         this.nextPositions = nextPositions;
